move filter in auth-client to directory
