+++ account-service/src/main/java/com/chatwave/accountservice/config/AppConfig.java
@@ -1,7 +1,7 @@
 package com.chatwave.accountservice.config;
 
 import com.chatwave.accountservice.domain.AccountMapper;
-import com.chatwave.authclient.UserAuthFilter;
+import com.chatwave.authclient.filter.UserAuthFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
